requirements.txt
groq
langchain-groq
langchain-openai
openai
presidio-analyzer
presidio-anonymizer
spacy
langgraph
langchain
langchain-community
langchain-astradb
astrapy
cassio
langchain-huggingface
sentence-transformers
fastapi
uvicorn
pydantic>=2.0.0
python-dotenv
pytest
pytest-asyncio
requests
streamlit
langgraph-checkpoint-cassandra
cassandra-driver
langchain-huggingface
huggingface_hub
langchain-google-genai
google-generativeai